Add comments 	modified:   data/loader.ipynb 	modified:   data/omicDC.py 	deleted:    loader.ipynb 	deleted:    omicDC.py
